Implement Army system and rudimentary combat (#8)
The trivial combat implementation calculates a win_pct as the battle simulation and rolls a random number to determine the battle result.

The major missing piece in playability still is a menu to replenish/recruit troops. That and there's a small inaccuracy in not checking regional neighbor status before allowing attacks.

After that it's just a matter of playtesting.