Added a very low-effort tutorial README
In game wall of text poorly explaining the game mechanics.

That's pretty much it. There's a handful of things that don't quite work right and a lot of things that are just lazy and badly designed for the sake of expediency but those will be improved in future versions.

Exported to windows and macos versions.