woops added too much
This includes basic assets for the battle scene and combat units, the battle and combat unit scenes themselves, and the logic for targeted combat abilities, ability effects, and effect ranges.

There are a few notable broken and missing features for getting it to feel doable, most notably ability targets are not respected, and the combat unit model updates aren't reflected in the UI yet, but it's not too far.