Implemented a animated solver
At current animation speeds it takes 23 minutes to solve the puzzle. This can be cranked up but looks progressively more weird.

Next I want to implement a faster in-memory solver that finds the shortest path and then either animate that or use it as the basis for a hint system.