Implement a path-finding algorithm for solution
This allows for a fast and slow solving modes.
This also adds a blocking layer to prevent controlling the
grid while the solver is running.
Also added a Deque implementation to try to optimize for double-ended access, but at this queue scale it's not an appreciable optimization.

Next up the main objective is to build a hint system. The idea will be to adapt the BFS algorithm to find all-paths to a final configuration and then allow the user to request a hint after all the paths have been computed.

This will require rebuilding the underlying data structures to uncouple it from the scene tree. This will simplify some things (like making the background thread processing thread-safe by default) but is likely a major rewrite.