Add hint system
This is basically in a working state.
I'd like to be able to shuffle the board state and have more board sizes, etc, but I think we should wrap this up.

I do want to add a little bit of a bow on it though. I think a saving system, start menu, and some music with volume controls at the last bits I want to add.