Sliding puzzle 4 (#16)
* Refactored tiles and grids to decouple state

This decouples the games state from the nodes so that we can more easily operate on a duplicate of the data in the background without manipulating the actually representation the player is interfacing with.

* almost have a working hint system

* Fix all path solver and add hint graphics

Everything should hopefully be working now, so we just need to store add the hint button and secret solution password.

* Add hint system

This is basically in a working state.
I'd like to be able to shuffle the board state and have more board sizes, etc, but I think we should wrap this up.

I do want to add a little bit of a bow on it though. I think a saving system, start menu, and some music with volume controls at the last bits I want to add.